requirements.txt
streamlit==1.28.1
pytesseract==0.3.10
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
paddleocr==2.7.0.3
paddle==2.5.1
requests==2.31.0
python-dotenv==1.0.0

